// src/interfaces/Vehicle.ts
export interface Vehicle {
  vehicle_name: string
  image: string
  model: string
  manufacturer: string
  cost_in_credits: string[]
  length: string[]
  max_atmosphere_speed: number
  crew: number
  passengers: number
  cargo_capacity: number
  consumables: string
  vehicle_class: number
  pilots: number
}